Security, Redirect, and Styling
Progress made on styling on the Home page. Updated colors on other existing pages. I also added an alert for when a user goes to submit a review when there is no movie or review text. User also cannot go to certain pages when they are not logged in, logged in, or are not an admin and will be redirected to the correct pages.