Edit/Delete Reviews works, Average Rating of Movie Kinda works
User can now edit and delete their reviews and will now see the average rating of all reviews for the movie that is searched up on the home page. However, the average rating will not update when the user submits their review and will only update when the page is refreshed and the movie is searched once more from the default state. Will have to fix that minor bug.